lesson-6-creating-web-search-agents-with-python/demo/requirements.txt
# Requirements for Sports Analyst Agent with External APIs Demo
# Core dependencies
python-dotenv==1.0.0
pydantic==2.5.0
semantic-kernel==0.3.0

# Azure AI Agents with Bing Search grounding
azure-ai-projects==1.0.0
azure-ai-agents==1.2.0b3
azure-identity

# Azure OpenAI dependencies
openai==1.3.0
azure-ai-formrecognizer==3.3.0
azure-ai-textanalytics==5.3.0
azure-cognitiveservices-speech==1.34.0

# External API dependencies
requests==2.31.0
urllib3==2.0.7
httpx==0.25.2

# Data processing (Python 3.13 compatible)
pandas>=2.3.1
numpy>=2.3.2

# Caching and performance
cachetools==5.3.2

# Logging and utilities
structlog==23.2.0
